% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasterize.R, R/package.R
\docType{package}
\name{fasterize}
\alias{fasterize}
\alias{fasterize}
\alias{fasterize-package}
\title{Rasterize an sf object of polygons}
\usage{
fasterize(polygons, raster, field = NULL, fun = "sum", background = NA)
}
\arguments{
\item{polygons}{an [sf::sf()] object with a geometry column of POLYGON and/or
MULTIPOLYGON objects.}

\item{raster}{A raster template object. Can be created with [raster::raster()].
The fasterize package provides a method to create a raster object from
an sf object.}

\item{field}{A numeric value or vector or the name of a column in `polygons`,
providing a value for each of the polygons rasterized.}

\item{fun}{The name of a function by which to combine overlapping polygons.
Currently only takes "sum", but future versions will provide other
high-performance compiled functions as well as the ability to pass arbitrary
R functions}

\item{background}{numeric. Value to put in the cells that are not covered by any of the features of x. Default is NA}
}
\value{
A raster of the same size, extent, resolution and projection as the
provided raster template.
}
\description{
This is a high-performance replacement for [raster::rasterize()].

Fast sf-to-raster conversion
}
\examples{
library(sf)
library(fasterize)
p1 <- rbind(c(-180,-20), c(-140,55), c(10, 0), c(-140,-60), c(-180,-20))
hole <- rbind(c(-150,-20), c(-100,-10), c(-110,20), c(-150,-20))
p1 <- list(p1, hole)
p2 <- list(rbind(c(-10,0), c(140,60), c(160,0), c(140,-55), c(-10,0)))
p3 <- list(rbind(c(-125,0), c(0,60), c(40,5), c(15,-45), c(-125,0)))
pols <- st_sf(value = rep(1,3),
              geometry = st_sfc(lapply(list(p1, p2, p3), st_polygon)))
r <- raster(pols, res = 1)
r <- fasterize(pols, r, field = "value", fun="sum")
plot(r)
}
