% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-methods.R
\name{raster,sf-method}
\alias{raster,sf-method}
\alias{raster}
\title{Create a raster from an sf object}
\usage{
\S4method{raster}{sf}(x, origin, ...)
}
\arguments{
\item{x}{an sf object}

\item{origin}{the origin point of the output raster}

\item{...}{Additional arguments, see \code{\link[raster:raster]{raster::raster()}} for more details.}
}
\description{
See \code{\link[raster:raster]{raster::raster()}} for more details.
}
