% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-methods.R
\name{raster}
\alias{raster}
\title{raster::plot}
\description{
re-exported \code{raster} method from \code{raster} package to avoid failing
when trying to plot raster objects without having \code{raster} loaded
first.
}
\keyword{internal}
