% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{fasterize-package}
\alias{fasterize-package}
\title{fasterize: Fast Polygon to Raster Conversion}
\description{
Provides a drop-in replacement for rasterize() from the 'raster' package that takes polygon vector or data frame objects, and is much faster. There is support for the main options provided by the rasterize() function, including setting the field used and background value, and options for aggregating multi-layer rasters. Uses the scan line algorithm attributed to Wylie et al. (1967) \doi{10.1145/1465611.1465619}.

Fast polygons-to-raster conversion
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecohealthalliance/fasterize}
  \item Report bugs at \url{https://github.com/ecohealthalliance/fasterize/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID}) [contributor]

Authors:
\itemize{
  \item Noam Ross \email{ross@ecohealthalliance.org} (\href{https://orcid.org/0000-0002-2136-0000}{ORCID}) (Original author)
}

Other contributors:
\itemize{
  \item Jeroen Ooms [contributor]
  \item Antoine Stevens [contributor]
  \item EcoHealth Alliance [copyright holder]
  \item USAID PREDICT [funder]
}

}
\keyword{internal}
